Fixed broken tests. Added combineReducers test for when input is a function.
