+++ lib/seedux/src/seeduxAssembler.js
@@ -1,11 +1,11 @@
 const parsedCodeObj = {};
 parsedCodeObj.uiResources = {};
 const uiHeadNode = new Node('Containers');
-uiHeadNode.children = [];
-
 const acHeadNode = new Node('Action Creators');
+uiHeadNode.children = [];
 acHeadNode.children = [];
 const actionMap = {};
+
 
 /**
  * Functional Node constructor used to assemble hierarchical tree objects.
@@ -124,6 +124,8 @@ function assembleActionCreatorNodes(actionCreatorNames, actionTypes) {
   parsedCodeObj.actionMap = actionMap;
   return acHeadNode;
 };
+
+// resetACHeadNode is used purely for testing purposes to assure acHeadNode does not persist data between tests
 
 function resetACHeadNode() {
   return acHeadNode.children = [];
