+++ lib/seedux/src/seeduxExtractor.js
@@ -94,7 +94,6 @@ function reducersExtractor(reducers) {
   const reducerNames = Object.keys(reducers);
   const strReducers = JSON.stringify(reducers);
   let reducerCases;
-  console.log('found reducers: ', reducers, reducerNames);
   if (strReducers.includes('switch')) {
     reducerCases = strReducers.split('case').slice(1).map(cases => cases.replace(/[^\w*:.]/g, '')).map(cases => cases = cases.slice(0, cases.indexOf(':')));
   }
@@ -126,13 +125,11 @@ function reducersExtractor(reducers) {
 */
 
 function actionCreatorsExtractor(actionCreators) {
-  console.log('Got actionCreators: ', actionCreators);
   const actionCreatorNames = Object.keys(actionCreators);
   const strActionCreators = JSON.stringify(actionCreators);
   let actionTypes;
 
   actionTypes = strActionCreators.split('type:').slice(1).map(subStr => subStr.slice(1, subStr.indexOf('\\')).replace(/[^\w*.,]/g, ''));
-  console.log(actionTypes);
   if (/\./.test(actionTypes)) {
     actionTypes = actionTypes.map(subStr => subStr.slice(subStr.indexOf('.') + 1));
   }
